Refactoring POE page -Moved title, filterform, main to poe page -matched other page functions for the site -changed how values passed to kanban element -added default page to poe route
